backend_app/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
PyMySQL==1.1.0
SQLAlchemy==2.0.28
cryptography==42.0.5
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-Cors==4.0.0
email-validator==2.1.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.28
PyMySQL==1.1.0
pydantic==2.5.1
mysql-connector-python==8.0.33 